move pill style function to util
