+++ src/utils/analysis.js
@@ -2,6 +2,33 @@ import { capitalize } from "./string";
 
 const ALL_SOS = ["SO1", "SO2", "SO3", "SO4", "SO5", "SO6"];
 
+export function getPillStyle(score) {
+  const bounds = {
+    bad: {
+      min: 0.0,
+      style: "badge-danger",
+    },
+    okay: {
+      min: 2.0,
+      style: "badge-warning",
+    },
+    good: {
+      min: 3.0,
+      style: "badge-success",
+    },
+  };
+  const baseStyle = "badge badge-pill";
+
+  if (score >= bounds.good.min) {
+    return `${baseStyle} ${bounds.good.style}`;
+  }
+  if (score >= bounds.okay.min && score < bounds.good.min) {
+    return `${baseStyle} ${bounds.okay.style}`;
+  } else {
+    return `${baseStyle} ${bounds.bad.style}`;
+  }
+}
+
 // Takes array of duplicate SO scores and returns array of averages for each SO
 function averageSOs(arr) {
   return arr.reduce((newData, current) => {
